Fixed:   BMC# 2043 - call history information is not saved...
BMC# 2043 - call history information is not saved for the first
            time that dialer is launched

Details: The first event added to the QSettings history cache was
failing to call endGroup(), causing the initial entry to be added
with extra data in the key.  Next time dialer is launched and we
read this entry, it will be invalid and result in "Unknown" being
shown in the Recent calls log.  Solved by adding endGroup() calls
prior to returning under empty cache conditions.

Signed-off-by: Shane Bryan <shane.bryan@linux.intel.com>